Updates following feedback from PR review.
