+++ core/src/main/kotlin/net/corda/core/internal/cordapp/ManifestUtils.kt
@@ -1,30 +1,7 @@
 package net.corda.core.internal.cordapp
 
 import net.corda.core.utilities.loggerFor
-import java.util.jar.Attributes
 import java.util.jar.Manifest
-
-fun createTestManifest(name: String, version: String, vendor: String, licence: String, targetVersion: Int): Manifest {
-    val manifest = Manifest()
-
-    // Mandatory manifest attribute. If not present, all other entries are silently skipped.
-    manifest.mainAttributes[Attributes.Name.MANIFEST_VERSION] = "1.0"
-
-    manifest[CORDAPP_CONTRACT_NAME] = name
-    manifest[CORDAPP_CONTRACT_VERSION] = version
-    manifest[CORDAPP_CONTRACT_VENDOR] = vendor
-    manifest[CORDAPP_CONTRACT_LICENCE] = licence
-
-    manifest[CORDAPP_WORKFLOW_NAME] = name
-    manifest[CORDAPP_WORKFLOW_VERSION] = version
-    manifest[CORDAPP_WORKFLOW_VENDOR] = vendor
-    manifest[CORDAPP_WORKFLOW_LICENCE] = licence
-
-    manifest[TARGET_PLATFORM_VERSION] = targetVersion.toString()
-    manifest[MIN_PLATFORM_VERSION] = "1"
-
-    return manifest
-}
 
 operator fun Manifest.set(key: String, value: String): String? {
     return mainAttributes.putValue(key, value)
@@ -38,8 +15,8 @@ operator fun Manifest.get(key: String): String? = mainAttributes.getValue(key)
 
 val Manifest.targetPlatformVersion: Int
     get() {
-        val minPlatformVersion = mainAttributes.getValue(MIN_PLATFORM_VERSION)?.toInt() ?: 1
-        return mainAttributes.getValue(TARGET_PLATFORM_VERSION)?.toInt() ?: minPlatformVersion
+        val minPlatformVersion = this[MIN_PLATFORM_VERSION]?.toIntOrNull() ?: 1
+        return this[TARGET_PLATFORM_VERSION]?.toIntOrNull() ?: minPlatformVersion
     }
 
 fun Manifest.toCordappInfo(defaultName: String): CordappInfo {
@@ -54,7 +31,7 @@ fun Manifest.toCordappInfo(defaultName: String): CordappInfo {
     if (this[CORDAPP_CONTRACT_NAME] != null) {
         val name = this[CORDAPP_CONTRACT_NAME] ?: defaultName
         val version = try {
-            Integer.valueOf(this[CORDAPP_CONTRACT_VERSION])
+            this[CORDAPP_CONTRACT_VERSION]?.toIntOrNull()
         } catch (nfe: NumberFormatException) {
             log.warn("Invalid version identifier ${this[CORDAPP_CONTRACT_VERSION]}. Defaulting to $DEFAULT_CORDAPP_VERSION")
             DEFAULT_CORDAPP_VERSION
@@ -74,7 +51,7 @@ fun Manifest.toCordappInfo(defaultName: String): CordappInfo {
     if (this[CORDAPP_WORKFLOW_NAME] != null) {
         val name = this[CORDAPP_WORKFLOW_NAME] ?: defaultName
         val version = try {
-            Integer.valueOf(this[CORDAPP_WORKFLOW_VERSION])
+            this[CORDAPP_WORKFLOW_VERSION]?.toIntOrNull()
         } catch (nfe: NumberFormatException) {
             log.warn("Invalid version identifier ${this[CORDAPP_CONTRACT_VERSION]}. Defaulting to $DEFAULT_CORDAPP_VERSION")
             DEFAULT_CORDAPP_VERSION
